test(ui): mount the providers page through its production component
Extract ProvidersTabContent into its own module, mirroring the sibling
provider-groups-content, and render that from the harness instead of
re-wiring the view's props by hand. The component owns the data path, so
scanConfigs and scanConfigStatus now reach the view exactly as they do in
production and the harness can no longer drift from it.

Opt the component out of the React Compiler: an async server component
renders once per request, so there is nothing to memoize, and the injected
useMemoCache leaves it uncallable outside a render.

Also narrow connectionCallCount to private, since only the harness's own
wait helper reads it.
